Split ir module
